+++ src/Foundation/Support/helpers.php
@@ -14,11 +14,19 @@ use Vanilo\MasterProduct\Contracts\MasterProductVariant;
  */
 function format_price($price, string $currency = null)
 {
-    return sprintf(
+    $result = sprintf(
         config('vanilo.foundation.currency.format'),
         $price,
         $currency ?? config('vanilo.foundation.currency.sign')
     );
+
+    if (is_string($decimalSeparator = config('vanilo.foundation.currency.decimal_separator')) && 1 === strlen($decimalSeparator)) {
+        $phpDecimalSeparator = localeconv()['decimal_point'] ?? '.';
+
+        $result = str_replace($phpDecimalSeparator, $decimalSeparator, $result);
+    }
+
+    return $result;
 }
 
 function is_master_product(object $product): bool